chore: utils is updated
