+++ frontend/src/store/actions/trace/parseFilter/filterToFetchData.ts
@@ -1,0 +1,37 @@
+import { TraceFilterEnum, TraceReducer } from 'types/reducer/trace';
+import { ParsedUrl } from '../util';
+
+export const parseFilterToFetchData = (
+	query: string,
+	stateTraceFilterData: TraceReducer['filterToFetchData'],
+): ParsedUrl<TraceFilterEnum[]> => {
+	const url = new URLSearchParams(query);
+
+	let filterToFetchData: TraceFilterEnum[] = [];
+
+	const selected = url.get('filterToFetchData');
+
+	if (selected) {
+		try {
+			const parsedValue = JSON.parse(decodeURIComponent(selected));
+
+			if (Array.isArray(parsedValue)) {
+				filterToFetchData.push(...parsedValue);
+			}
+		} catch (error) {
+			console.log('error while parsing json');
+		}
+	}
+
+	if (selected) {
+		return {
+			currentValue: filterToFetchData,
+			urlValue: filterToFetchData,
+		};
+	}
+
+	return {
+		currentValue: stateTraceFilterData,
+		urlValue: filterToFetchData,
+	};
+};
