+++ frontend/src/store/actions/trace/parseFilter/selectedFilter.ts
@@ -1,0 +1,41 @@
+import { TraceFilterEnum, TraceReducer } from 'types/reducer/trace';
+import { isTraceFilterEnum, ParsedUrl } from '../util';
+
+export const parseSelectedFilter = (
+	query: string,
+	selectedFilter: TraceReducer['selectedFilter'],
+	isUserSelected = false,
+): ParsedUrl<Map<TraceFilterEnum, string[]>> => {
+	const url = new URLSearchParams(query);
+
+	const filters = new Map<TraceFilterEnum, string[]>();
+
+	const selected = url.get(isUserSelected ? 'selected' : 'userSelectedFilter');
+
+	if (selected) {
+		try {
+			const parsedValue = JSON.parse(decodeURIComponent(selected));
+			if (typeof parsedValue === 'object') {
+				Object.keys(parsedValue).forEach((e) => {
+					if (isTraceFilterEnum(e)) {
+						filters.set(e, parsedValue[e]);
+					}
+				});
+			}
+		} catch (error) {
+			// if the parsing error happens
+		}
+	}
+
+	if (selected) {
+		return {
+			urlValue: filters,
+			currentValue: filters,
+		};
+	}
+
+	return {
+		urlValue: filters,
+		currentValue: selectedFilter,
+	};
+};
